chore: move mutation to an Operations folder because it's not a component
