+++ src/v2/Apps/Consign/OfferDetail/Components/ResponseForm.tsx
@@ -2,7 +2,7 @@ import React, { useState } from "react"
 import { Form, Formik } from "formik"
 import { BorderedRadio, Button, RadioGroup, Text } from "@artsy/palette"
 import { useSystemContext } from "v2/Artsy"
-import { CreateOfferResponse } from "./CreateOfferResponse"
+import { CreateOfferResponse } from "../Operations/CreateOfferResponse"
 import { ResponseForm_offer } from "v2/__generated__/ResponseForm_offer.graphql"
 import { createFragmentContainer, graphql } from "react-relay"
 import { CreateOfferResponseMutationInput } from "v2/__generated__/CreateOfferResponseMutation.graphql"
