+++ src/v2/Apps/Consign/OfferDetail/Operations/CreateOfferResponse.ts
